Move the planned-stations client out of storage/
storage/ holds the visit-persistence abstraction, which this module is not
part of, so the development-plan client now sits beside the other plan modules
as src/frontend/planned-stations.ts.
